restructuring and new UI transitions
