+++ src/background/camera-controller/MouseCameraController.test.tsx
