+++ src/background/camera-controller/MouseCameraController.tsx
