+++ src/content-ui/OverlayPage.css
@@ -16,21 +16,22 @@
   left: 0;
   width: 100%;
   height: 100%;
-  backdrop-filter: blur(5px);
-  background: rgba(0, 0, 0, 0.1);
+  background: #000000;
   cursor: pointer;
+  opacity: 0;
+  animation: fadeInBackground 0.3s ease-out 0.4s forwards;
 }
 
 .overlay-content {
   position: relative;
   width: 100vw;
   height: 100vh;
-  background: rgba(5, 5, 5, 0.3);
-  backdrop-filter: blur(15px);
+  background: #000000;
   overflow: hidden;
   display: flex;
   flex-direction: column;
-  opacity: 1;
+  opacity: 0;
+  animation: fadeInContent 0.4s ease-out 0.6s forwards;
 }
 
 
@@ -92,6 +93,24 @@
 }
 
 /* Overlay fade-in animation */
+@keyframes fadeInBackground {
+  from {
+    opacity: 0;
+  }
+  to {
+    opacity: 1;
+  }
+}
+
+@keyframes fadeInContent {
+  from {
+    opacity: 0;
+  }
+  to {
+    opacity: 1;
+  }
+}
+
 @keyframes fadeInOverlay {
   from {
     opacity: 0;
