+++ src/content-ui/pages/about-page/AboutSection.css
@@ -1,0 +1,4 @@
+/* About Section Styles */
+.about-content {
+  max-width: 800px;
+}
