+++ src/navigation-ui/GalaxyDebugOverlay.css
